pretty much done with tutorial 3
There is a wierd thing going on when I try to solve the offset correction problem with values slighly off the normal [1 0 0 0], and have an upper bound on the input of: ub = 0.43 or lower...
ipopt returns "infeasible problem detected", but it does't make sense for the problem to be in feasible...
The built-in sqp algorithm also fails in the same scenarios as ipopt. Must be an issue with the problem formulation itself. (crap...)